+++ android/app/src/main/kotlin/com/example/jsend/MainActivity.kt
@@ -1,6 +1,0 @@
-package com.example.jsend
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}