+++ lib/main.dart
@@ -1,19 +1,37 @@
-// import 'api_request.dart';
-// import 'jsend.dart';
+import 'dart:math';
 
-// void main() async {
-//   APIRequest.base = 'http://localhost:5000/api/ads/';
-//   var apiRequest = APIRequest(
-//     // path: 'auth/register',
-//     url: 'http://localhost:5000/api/ads/auth/register',
-//     payload: {'name': 'Arjun', 'email': 'arju@gmail.com'},
-//     method: 'POST',
-//   );
+import 'package:jsend/api_request.dart';
+import 'package:jsend/jsend.dart';
+import 'package:jsend/remote_resource.dart';
 
-//   //testing apirequest pending
+void main() async {
+  APIRequest.base = 'http://localhost:5000/api/dalle/';
+  var resource = RemoteResource('addons');
+  // var all = await resource.getAll();
 
-//   var httpResponse = await apiRequest.send();
-//   var jsonResp = jsendResponse(httpResponse);
-//   print(httpResponse.body);
-//   print(jsonResp);
-// }
+  // print(all);
+  //
+  var jr = jsendResponse.fromAPIRequest(
+    APIRequest(path: 'quantities'),
+    onError: (jsendResponse r) {
+      print('1');
+    },
+    onFail: defaultStatusHandler,
+    statusHandlers: JsendStatusHandlers(forError: (jsr) {
+      print('2');
+      // print(jsr);
+    }),
+  );
+
+  var rand = Random();
+
+  // var payload = {
+  //   'name': 'Addon ' + rand.nextInt(200).toString(),
+  // };
+
+  // await resource.createItem(payload);
+
+  // print(await resource.getAll());
+
+  // print('last');
+}